requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
scikit-learn
opencv-python
flask
pillow
gunicorn
